refactor(training): move training orchestration into training package (Closes #54)
