+++ src/baseline.py
@@ -1,7 +1,7 @@
-"""Compatibility wrapper for the baseline model pipeline."""
+"""Compatibility wrapper for baseline training."""
 
-from src.models.baseline import *  # noqa: F401,F403
-from src.models.baseline import main
+from src.training.baseline import *  # noqa: F401,F403
+from src.training.baseline import main
 
 
 if __name__ == "__main__":
